refactor: move default toToken logic to util
